+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -1,12 +1,12 @@
 package kitchenpos.menu.application;
 
-import static kitchenpos.application.exception.ExceptionType.NOT_FOUND_MENU_GROUP_EXCEPTION;
-import static kitchenpos.application.exception.ExceptionType.NOT_FOUND_PRODUCT_EXCEPTION;
+import static kitchenpos.exception.ExceptionType.NOT_FOUND_MENU_GROUP_EXCEPTION;
+import static kitchenpos.exception.ExceptionType.NOT_FOUND_PRODUCT_EXCEPTION;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
-import kitchenpos.application.exception.CustomIllegalArgumentException;
+import kitchenpos.exception.CustomIllegalArgumentException;
 import kitchenpos.dao.JpaMenuGroupRepository;
 import kitchenpos.menu.domain.JpaMenuRepository;
 import kitchenpos.product.domain.JpaProductRepository;
