+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.domain;
 
-import static kitchenpos.application.exception.ExceptionType.INVALID_MENU_PRODUCT_EXCEPTION;
+import static kitchenpos.exception.ExceptionType.INVALID_MENU_PRODUCT_EXCEPTION;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
@@ -15,7 +15,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
-import kitchenpos.application.exception.CustomIllegalArgumentException;
+import kitchenpos.exception.CustomIllegalArgumentException;
 import kitchenpos.domain.Price;
 
 @Entity
